# Optimize PathUtilities.Sanitize() and DeterministicGuid.Create()