--- a/src/NServiceBus.Core/Utils/DeterministicGuid.cs
+++ b/src/NServiceBus.Core/Utils/DeterministicGuid.cs
@@ -6,16 +6,19 @@
 
     static class DeterministicGuid
     {
-        public static Guid Create(params object[] data)
+        public static Guid Create(string data1, string data2) => Create($"{data1}{data2}");
+
+        public static Guid Create(string data)
         {
             // use MD5 hash to get a 16-byte hash of the string
-            using (var provider = MD5.Create())
-            {
-                var inputBytes = Encoding.Default.GetBytes(string.Concat(data));
-                var hashBytes = provider.ComputeHash(inputBytes);
-                // generate a guid from the hash:
-                return new Guid(hashBytes);
-            }
+            var inputBytes = Encoding.Default.GetBytes(data);
+
+            Span<byte> hashBytes = stackalloc byte[16];
+
+            _ = MD5.HashData(inputBytes, hashBytes);
+
+            // generate a guid from the hash:
+            return new Guid(hashBytes);
         }
     }
 }