@danielmarbach 2023-09-13
@stevenaw Thanks for the contribution and taking the time to even validate your changes with Benchmark.NET. That's a wonderful quality PR. Kudos!

One thing to note. Those utilities are not really ever used on the hot path. They are mostly only used on the endpoint creation layer or in the learning persistence, which is not really performance sensitive. That being said, since you invested already some time and those are "utils" classes, I'm willing to say those get used or copied and pasted under the assumption it is good enough to be used anywhere. 

I took the commits you provided and did an additional round of improvements on the deterministic guid this the following PR

https://github.com/Particular/NServiceBus/pull/6848

The PR is currently not passing because we have an analyzer problem with the latest RC1 updates shipped yesterday.