@danielmarbach 2023-09-19
I have merged https://github.com/Particular/NServiceBus/pull/6848. Thank you! Could you reach out to daniel dot marbach at particular dot net? I would love to give you a small gift of appreciation for your first contribution. Thanks